+++ src/main/java/dashboard/controllers/product/ProductCategoryController.java
@@ -8,15 +8,13 @@ import dashboard.enums.EntityStatus;
 import dashboard.exceptions.customs.ResourceNotFoundException;
 import dashboard.services.PusherService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import dashboard.entities.ProductCategory;
+import dashboard.entities.product.ProductCategory;
 import dashboard.services.ProductCategoryService;
 
 @RestController
