+++ src/main/java/dashboard/controllers/product/ProductTypeGroupController.java
@@ -1,16 +1,13 @@
 package dashboard.controllers.product;
 
 import dashboard.constants.PusherConstants;
-import dashboard.entities.ProductCategory;
-import dashboard.entities.ProductTypeGroup;
+import dashboard.entities.product.ProductTypeGroup;
 import dashboard.enums.EntityStatus;
 import dashboard.exceptions.customs.ResourceNotFoundException;
 import dashboard.generics.MultipleExecute;
-import dashboard.repositories.ProductTypeGroupRepository;
 import dashboard.services.ProductTypeGroupService;
 import dashboard.services.PusherService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
@@ -18,8 +15,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/product/type_group")
