+++ src/main/java/dashboard/entities/post/PostType.java
@@ -1,5 +1,6 @@
-package dashboard.entities;
+package dashboard.entities.post;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import dashboard.entities.base.BaseEntity;
 import dashboard.enums.EntityStatus;
@@ -10,42 +11,40 @@ import javax.validation.constraints.NotNull;
 import java.io.Serializable;
 
 @Entity
-@Table(name = "product_brand")
+@Table( name = "post_type")
 @EntityListeners(AuditingEntityListener.class)
-public class ProductBrand extends BaseEntity implements Serializable {
+@JsonIgnoreProperties(value = {"createDate", "updateDate", "deleleDate"},
+        allowGetters = true)
+public class PostType extends BaseEntity implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "product_brand_id")
-    @JsonProperty("product_brand_id")
-    private Long productBrandId;
+    @Column(name = "post_type_id")
+    @JsonProperty("post_type_id")
+    private Long  postTypeId;
 
     @NotNull(message = "Name is not null")
     @Column(name = "name")
     @JsonProperty("name")
     private String name;
 
-    @NotNull(message = "Slug name is not null")
+    @NotNull(message = "Slug_name is not null")
     @Column(name = "slug_name")
     @JsonProperty("slug_name")
     private String slugName;
-
-    @Column(name = "image")
-    @JsonProperty("image")
-    private String image;
 
     @Column(name = "status")
     @Enumerated(EnumType.STRING)
     private EntityStatus status;
 
-    public Long getProductBrandID() {
-        return productBrandId;
+    public Long getPostTypeId() {
+        return postTypeId;
     }
 
-    public void setProductBrandID(Long productBrandID) {
-        this.productBrandId = productBrandID;
+    public void setPostTypeId(Long postTypeId) {
+        this.postTypeId = postTypeId;
     }
 
     public String getName() {
@@ -62,14 +61,6 @@ public class ProductBrand extends BaseEntity implements Serializable {
 
     public void setSlugName(String slugName) {
         this.slugName = slugName;
-    }
-
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
     }
 
     public EntityStatus getStatus() {
