+++ src/main/java/dashboard/repositories/ProductBrandRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import dashboard.entities.ProductBrand;
+import dashboard.entities.product.ProductBrand;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
