+++ src/main/java/dashboard/repositories/ProductCategoryRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import dashboard.entities.ProductCategory;
+import dashboard.entities.product.ProductCategory;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
