+++ src/main/java/dashboard/services/PostTypeService.java
@@ -1,13 +1,11 @@
 package dashboard.services;
 
-import dashboard.entities.PostType;
-import dashboard.entities.ProductBrand;
+import dashboard.entities.post.PostType;
 import dashboard.enums.EntityStatus;
 import dashboard.exceptions.customs.ResourceNotFoundException;
 import dashboard.generics.ListEntityResponse;
 import org.springframework.data.domain.Pageable;
 
-import java.util.Date;
 import java.util.List;
 
 public interface PostTypeService {
