+++ src/main/java/dashboard/services/implement/ProductBrandSerivceImpl.java
@@ -1,7 +1,6 @@
 package dashboard.services.implement;
 
-import dashboard.entities.ProductBrand;
-import dashboard.entities.ProductCategory;
+import dashboard.entities.product.ProductBrand;
 import dashboard.enums.EntityStatus;
 import dashboard.exceptions.customs.ResourceNotFoundException;
 import dashboard.generics.ListEntityResponse;
