+++ src/main/java/dashboard/services/implement/ProductCategorySerivceImpl.java
@@ -8,7 +8,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import dashboard.entities.ProductCategory;
+import dashboard.entities.product.ProductCategory;
 import dashboard.repositories.ProductCategoryRepository;
 import dashboard.services.ProductCategoryService;
